# Improvements to KafkaSource adapter