## Proposed Changes

  * Added memory pool to reuse memory for created events
  * Removed useless marshalling/unmarshalling of json