@matzew 2019-09-17
/test pull-knative-eventing-contrib-unit-tests