@slinkydeveloper 2019-09-17
Before:

```
BenchmarkHandle/Baseline-4              10000000               255 ns/op             163 B/op          2 allocs/op
BenchmarkHandle/Benchmark_Handle-4         10000            235736 ns/op           19476 B/op        297 allocs/op
```

Now:

```
BenchmarkHandle/Baseline-4              10000000               265 ns/op             163 B/op          2 allocs/op
BenchmarkHandle/Benchmark_Handle-4         10000            205500 ns/op           16158 B/op        258 allocs/op
```

Benchmark pr incoming